Code-Review-Agent/requirements.txt
tornado>=6.4
openai>=1.30
requests>=2.31
python-dotenv>=1.0
pylint>=3.0
bandit>=1.7
